UI working

+++ src/main/java/com/americancsm/gdpr/webassess/model/GDPRAssessmentResponse.java
@@ -9,6 +9,6 @@ public @Data class GDPRAssessmentResponse implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	private String result;
-	private Float score;
+	private Integer score;
 
 }
